git-svn-id: http://csat.googlecode.com/svn/trunk@38 b6245d9c-a44d-11dd-a9c7-49dec45ca499

+++ csat_src/VBO.cs
@@ -117,11 +117,11 @@ namespace CSat
         {
             GL.GenBuffers(1, out vertexID);
             GL.BindBuffer(BufferTarget.ArrayBuffer, vertexID);
-            GL.BufferData(BufferTarget.ArrayBuffer, (IntPtr)(vertSize * vertexSize), null, usage);
+            GL.BufferData(BufferTarget.ArrayBuffer, (IntPtr)(vertSize * vertexSize), (IntPtr)null, usage);
 
             GL.GenBuffers(1, out indexID);
             GL.BindBuffer(BufferTarget.ElementArrayBuffer, indexID);
-            GL.BufferData(BufferTarget.ElementArrayBuffer, (IntPtr)(indSize * sizeof(int)), null, usage);
+            GL.BufferData(BufferTarget.ElementArrayBuffer, (IntPtr)(indSize * sizeof(int)), (IntPtr)null, usage);
 
             numOfIndices = indSize;
